+++ R/utils.R
@@ -47,3 +47,7 @@ names(mnthqrt2) <- rep(1:4, each = 3)
 mnthqrt3 <- 1:12
 names(mnthqrt3) <- rep(1:4, each = 3)
 
+
+na_omit <- function(x, ...){
+    x[!is.na(x)]
+}